+++ OntraportApi/IOntraportTasks.cs
@@ -14,7 +14,7 @@ public interface IOntraportTasks : IOntraportBaseRead<ApiTask>
     /// <param name="dateDue">The date and time the task should be due.</param>
     /// <param name="status">The task's status.</param>
     /// <returns>An object containing updated fields.</returns>
-    Task<ApiTask> UpdateAsync(int taskId, int? owner = null, DateTimeOffset? dateDue = null, ApiTask.TaskStatus? status = null, CancellationToken cancellationToken = default);
+    Task<ApiTask> UpdateAsync(long taskId, long? owner = null, DateTimeOffset? dateDue = null, ApiTask.TaskStatus? status = null, CancellationToken cancellationToken = default);
 
     /// <summary>
     /// Assigns a task to one or more contacts.
@@ -46,5 +46,5 @@ public interface IOntraportTasks : IOntraportBaseRead<ApiTask>
     /// <param name="id">The task ID.</param>
     /// <param name="newTime">The date and time the task should be rescheduled for.</param>
     /// <returns></returns>
-    Task RescheduleAsync(int id, DateTimeOffset newTime, CancellationToken cancellationToken = default);
+    Task RescheduleAsync(long id, DateTimeOffset newTime, CancellationToken cancellationToken = default);
 }
